+++ jetbrains-core/src/software/aws/toolkits/jetbrains/core/gettingstarted/editor/GettingStartedPanel.kt
@@ -3,16 +3,327 @@
 
 package software.aws.toolkits.jetbrains.core.gettingstarted.editor
 
+import com.intellij.ide.ui.laf.darcula.ui.DarculaButtonUI
+import com.intellij.openapi.actionSystem.ActionManager
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.PlatformDataKeys
+import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.OpenFileDescriptor
+import com.intellij.openapi.project.DefaultProjectFactory
+import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
+import com.intellij.ui.IdeBorderFactory
+import com.intellij.ui.JBColor
+import com.intellij.ui.components.JBLabel
+import com.intellij.ui.dsl.builder.TopGap
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.util.ui.JBFont
+import com.intellij.util.ui.UIUtil
 import com.intellij.util.ui.components.BorderLayoutPanel
-import javax.swing.JLabel
+import icons.AwsIcons
+import software.aws.toolkits.jetbrains.services.caws.CawsEndpoints
+import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererConstants.CODEWHISPERER_LEARN_MORE_URI
+import software.aws.toolkits.jetbrains.ui.feedback.FeedbackDialog
+import software.aws.toolkits.resources.message
+import java.awt.Color
+import java.awt.Dimension
 
-class GettingStartedPanel : BorderLayoutPanel() {
+class GettingStartedPanel(private val project: Project) : BorderLayoutPanel() {
     init {
-        addToCenter(JLabel("Hello world"))
+        addToCenter(
+            panel {
+                indent {
+                    row {
+                        apply {
+                            icon(AwsIcons.Logos.AWS_SMILE_LARGE)
+                        }
+                        panel {
+                            row {
+                                label(message("aws.onboarding.getstarted.panel.title")).applyToComponent {
+                                    font = JBFont.h1().asBold()
+                                    foreground = PanelText.TITLE_TEXT_FONTCOLOR
+                                }
+                            }
+                            row {
+                                text(message("aws.onboarding.getstarted.panel.comment_link_doc"))
+                                text(message("aws.onboarding.getstarted.panel.comment_link_github"))
+                                text(message("aws.onboarding.getstarted.panel.share_feedback")) { hyperlinkEvent ->
+                                    val actionEvent = AnActionEvent.createFromInputEvent(
+                                        hyperlinkEvent.inputEvent,
+                                        PanelText.SHARE_FEEDBACK_LINK,
+                                        null
+                                    ) { if (PlatformDataKeys.PROJECT.`is`(it)) project else null }
+                                    ActionManager.getInstance().getAction("aws.toolkit.getstarted.shareFeedback").actionPerformed(actionEvent)
+                                }
+                            }
+                        }
+                    }.topGap(TopGap.MEDIUM)
+
+                    group(
+                        JBLabel(message("aws.onboarding.getstarted.panel.group_title"))
+                            .apply {
+                                font = JBFont.h2().asBold()
+                                foreground = PanelText.TITLE_TEXT_FONTCOLOR
+                            }
+                    ) {
+                        row {
+                            // CodeWhisperer panel
+                            cell(CodeWhispererPanel())
+                            // Resource Explorer Panel
+                            cell(ResourceExplorerPanel())
+                            // CodeCatalyst Panel
+                            cell(CodeCatalystPanel())
+                        }
+                    }
+
+                    collapsibleGroup(message("aws.onboarding.getstarted.panel.bottom_text_question")) {
+                        row {
+                            text(message("aws.onboarding.getstarted.panel.bottom_text"))
+                        }
+                        row {
+                            // CodeWhisperer auth bullets
+                            cell(PanelAuthBullets(message("codewhisperer.experiment")))
+//                // Resource Explorer panel auth bullets
+                            cell(PanelAuthBullets(message("aws.getstarted.resource.panel_title")))
+//                // CodeCatalyst panel auth bullets
+                            cell(PanelAuthBullets(message("caws.devtoolPanel.title")))
+                        }
+                    }
+                }
+            }
+        )
     }
+
+    class CodeCatalystPanel : BorderLayoutPanel() {
+        init {
+            addToCenter(
+                panel {
+                    indent {
+                        row {
+                            label(message("caws.devtoolPanel.title")).bold()
+                                .applyToComponent {
+                                    foreground = PanelText.TITLE_TEXT_FONTCOLOR
+                                }
+                        }
+
+                        row {
+                            panel {
+                                row {
+                                    text(message("caws.getstarted.panel.image"))
+                                }
+                            }
+                        }
+
+                        row {
+                            text(message("caws.getstarted.panel.description")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+
+                        row {
+                            browserLink(message("codewhisperer.gettingstarted.panel.learn_more"), CawsEndpoints.ConsoleFactory.baseUrl())
+                        }
+
+                        row {
+                            button(message("caws.getstarted.panel.login")) {}.applyToComponent {
+                                putClientProperty(DarculaButtonUI.DEFAULT_STYLE_KEY, true)
+                            }
+                        }
+
+                        row {
+                            label(message("caws.getstarted.panel.question.text")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+                        row {
+                            browserLink(message("caws.getstarted.panel.link_text"), CawsEndpoints.CAWS_SPACES_DOC)
+                        }
+                    }
+                }
+            )
+
+            border = IdeBorderFactory.createRoundedBorder().apply {
+                setColor(PanelText.TEXT_FONTCOLOR)
+                preferredSize = Dimension(PanelText.PANEL_WIDTH, PanelText.PANEL_HEIGHT)
+            }
+        }
+    }
+
+    class ResourceExplorerPanel : BorderLayoutPanel() {
+        init {
+            addToCenter(
+                panel {
+                    indent {
+                        row {
+                            label(message("aws.getstarted.resource.panel_title")).bold()
+                                .applyToComponent {
+                                    foreground = PanelText.TITLE_TEXT_FONTCOLOR
+                                }
+                        }
+                        row {
+                            panel {
+                                row {
+                                    text(message("caws.getstarted.panel.image"))
+                                }
+                            }
+                        }
+
+                        row {
+                            text(message("aws.getstarted.resource.panel_description")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+
+                        row {
+                            browserLink(
+                                message("codewhisperer.gettingstarted.panel.learn_more"),
+                                url = "https://docs.aws.amazon.com/toolkit-for-jetbrains/latest/userguide/working-with-aws.html"
+                            )
+                        }
+
+                        row {
+                            button(message("aws.onboarding.getstarted.panel.button_iam_login")) {}
+                                .applyToComponent {
+                                    putClientProperty(DarculaButtonUI.DEFAULT_STYLE_KEY, true)
+                                }
+                            topGap(TopGap.MEDIUM)
+                        }
+
+                        row {
+                            label(message("aws.getstarted.resource.panel_question_text")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+                        row {
+                            browserLink(message("aws.onboarding.getstarted.panel.signup_iam_text"), url = "https://aws.amazon.com/free/")
+                        }
+                    }
+                }
+            )
+
+            border = IdeBorderFactory.createRoundedBorder().apply {
+                setColor(PanelText.TEXT_FONTCOLOR)
+                preferredSize = Dimension(PanelText.PANEL_WIDTH, PanelText.PANEL_HEIGHT)
+            }
+        }
+    }
+
+    class CodeWhispererPanel : BorderLayoutPanel() {
+        init {
+            addToCenter(
+                panel {
+                    indent {
+                        row {
+                            label(message("codewhisperer.experiment")).bold()
+                                .applyToComponent { foreground = PanelText.TITLE_TEXT_FONTCOLOR }
+                        }
+                        row {
+                            panel {
+                                row {
+                                    text(message("caws.getstarted.panel.image"))
+                                }
+                            }
+                        }
+
+                        row {
+                            text(message("codewhisperer.gettingstarted.panel.comment")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+
+                        row {
+                            browserLink(message("codewhisperer.gettingstarted.panel.learn_more"), url = CODEWHISPERER_LEARN_MORE_URI)
+                        }
+
+                        row {
+                            button(message("codewhisperer.gettingstarted.panel.login_button")) {}
+                                .applyToComponent {
+                                    putClientProperty(DarculaButtonUI.DEFAULT_STYLE_KEY, true)
+                                }
+                            topGap(TopGap.SMALL)
+                        }
+
+                        row {
+                            label(message("codewhisperer.gettingstarted.panel.licence_comment")).applyToComponent { foreground = PanelText.TEXT_FONTCOLOR }
+                        }
+                        row {
+                            browserLink(message("aws.onboarding.getstarted.panel.login_with_iam"), url = CODEWHISPERER_LEARN_MORE_URI)
+                        }
+                    }
+                }
+            )
+
+            border = IdeBorderFactory.createRoundedBorder().apply {
+                setColor(PanelText.TEXT_FONTCOLOR)
+                preferredSize = Dimension(PanelText.PANEL_WIDTH, PanelText.PANEL_HEIGHT)
+            }
+        }
+    }
+
+    class PanelAuthBullets(private var panelTitle: String) : BorderLayoutPanel() {
+
+        val codeWhispererBulletsLists: List<BulletAuthPanel> = listOf(
+            BulletAuthPanel(PanelText.COMMIT_ICON, message("iam_identity_center.name"), message("aws.onboarding.getstarted.panel.idc_row_comment_text")),
+            BulletAuthPanel(PanelText.COMMIT_ICON, message("aws_builder_id.service_name"), "Need to insert tagline"),
+            BulletAuthPanel(PanelText.CANCEL_ICON, message("settings.credentials.iam"), message("aws.getstarted.auth.panel.notSupport_text"), false)
+        )
+
+        val resourceBulletsLists: List<BulletAuthPanel> = listOf(
+            BulletAuthPanel(PanelText.COMMIT_ICON, message("iam_identity_center.name"), message("aws.onboarding.getstarted.panel.idc_row_comment_text")),
+            BulletAuthPanel(PanelText.CANCEL_ICON, message("aws_builder_id.service_name"), "Need to insert tagline", false),
+            BulletAuthPanel(PanelText.COMMIT_ICON, message("settings.credentials.iam"), message("aws.getstarted.auth.panel.notSupport_text"))
+        )
+
+        val codeCatalystBulletsLists: List<BulletAuthPanel> = listOf(
+            BulletAuthPanel(PanelText.CANCEL_ICON, message("iam_identity_center.name"), message("aws.getstarted.auth.panel.notSupport_text"), false),
+            BulletAuthPanel(PanelText.COMMIT_ICON, message("aws_builder_id.service_name"), "Need to insert tagline"),
+            BulletAuthPanel(PanelText.CANCEL_ICON, message("settings.credentials.iam"), message("aws.getstarted.auth.panel.notSupport_text"), false)
+        )
+
+        init {
+
+            val serviceTitleMap: MutableMap<String, List<BulletAuthPanel>> = HashMap()
+            serviceTitleMap.put(message("codewhisperer.experiment"), codeWhispererBulletsLists)
+            serviceTitleMap.put(message("aws.getstarted.resource.panel_title"), resourceBulletsLists)
+            serviceTitleMap.put(message("caws.devtoolPanel.title"), codeCatalystBulletsLists)
+
+            addToCenter(
+                panel {
+                    indent {
+
+                        row {
+                            label(panelTitle).bold().applyToComponent { foreground = PanelText.TITLE_TEXT_FONTCOLOR }
+                        }
+
+                        for (bullet in serviceTitleMap.get(panelTitle)!!) {
+                            row {
+                                text(bullet.icon)
+                                panel {
+                                    row(bullet.titleName) {
+                                    }.rowComment(bullet.comment)
+                                        .enabled(bullet.enable)
+                                }
+                            }
+                        }
+                    }
+                }
+            )
+
+            border = IdeBorderFactory.createRoundedBorder().apply {
+                setColor(PanelText.TEXT_FONTCOLOR)
+                preferredSize = Dimension(PanelText.PANEL_WIDTH, PanelText.BULLET_PANEL_HEIGHT)
+            }
+        }
+    }
+
+    object PanelText {
+        const val SHARE_FEEDBACK_LINK = "FeedbackDialog"
+        const val COMMIT_ICON = "<icon src='AllIcons.General.InspectionsOK'/>&nbsp;"
+        const val CANCEL_ICON = "<icon src='AllIcons.CodeWithMe.CwmTerminate'/>&nbsp;"
+        val TEXT_FONTCOLOR = UIUtil.getLabelForeground()
+        val TITLE_TEXT_FONTCOLOR = JBColor(Color.WHITE, Color.WHITE)
+        const val PANEL_WIDTH = 300
+        const val PANEL_HEIGHT = 450
+        const val BULLET_PANEL_HEIGHT = 200
+    }
+
+    data class BulletAuthPanel(
+        val icon: String,
+        val titleName: String,
+        val comment: String,
+        val enable: Boolean = true
+    )
 
     companion object {
         fun openPanel(project: Project) = FileEditorManager.getInstance(project).openTextEditor(
@@ -22,5 +333,14 @@ class GettingStartedPanel : BorderLayoutPanel() {
             ),
             true
         )
+    }
+}
+
+class ShareFeedbackInGetStarted : DumbAwareAction() {
+
+    override fun actionPerformed(e: AnActionEvent) {
+        runInEdt {
+            FeedbackDialog(DefaultProjectFactory.getInstance().defaultProject).show()
+        }
     }
 }
