+++ resources/resources/software/aws/toolkits/resources/MessagesBundle.properties
@@ -110,7 +110,7 @@ aws.onboarding.getstarted.panel.bottom_text=1 or 2 sentence about why these are 
 aws.onboarding.getstarted.panel.button_iam_login=Authenticate with IAM
 aws.onboarding.getstarted.panel.comment_link_doc=<a href = "https://docs.aws.amazon.com/toolkit-for-jetbrains/latest/userguide/welcome.html">Documentation</a> | 
 aws.onboarding.getstarted.panel.comment_link_github=<a href="https://github.com/aws/aws-toolkit-jetbrains">Join us on github</a> |
-aws.onboarding.getstarted.panel.group_title=<html><h2 style="color:{0}"; "font-size= 14px"><b>Select a feature to setup authentication for</b></h2></html>
+aws.onboarding.getstarted.panel.group_title=Select a feature to setup authentication for
 aws.onboarding.getstarted.panel.idc_row_comment_text=Successor to AWS Single Sign-on 
 aws.onboarding.getstarted.panel.login_with_iam=Login with IAM Identity Center
 aws.onboarding.getstarted.panel.share_feedback=<a>Share Feedback</a>
@@ -229,6 +229,7 @@ caws.environment.status=Dev Environment is currently: {0}
 caws.environment.view_pricing=Learn more about Dev Environment pricing
 caws.free.tier.subscription.storage=Other storage options are not currently available for your space. These require upgrading the billing tier.
 caws.getstarted.panel.description=Spend more time coding and less <br>time managing development <br>environments.
+caws.getstarted.panel.image=<img src="https://cdn.svgator.com/images/2022/11/Animated-GIF-best-practices.gif" height="400" width="500">
 caws.getstarted.panel.link_text=Learn more about CodeCatalyst Spaces
 caws.getstarted.panel.login=Connect to CodeCatalyst
 caws.getstarted.panel.question.text=Don't have a CodeCatalyst Space?
